rename: custom_shell/commands.py -> command.py
